Move Metabot QB and Sidebar into query builder
