+++ frontend/src/metabase/query_builder/components/AIQuestionAnalysisSidebar/index.ts
