+++ frontend/src/metabase/query_builder/components/AIQuestionAnalysisSidebar/utils.ts
